@@ -1,48 +1,49 @@
-var precise = require("../lib/precise");
+import {default as assert} from "node:assert";
+import {precise} from "../dist/precise.esm.js";
 
-exports["start"] = {
-	setUp: function (done) {
+describe("Testing starting state", function () {
+	beforeEach(function () {
 		this.timer = precise();
-		done();
-	},
-	tests: function (test) {
-		test.expect(5);
-		test.strictEqual(this.timer.started.length, 0, "Should be '0'");
-		test.strictEqual(this.timer.stopped.length, 0, "Should be '0'");
+	});
+
+	it("It should be empty", function () {
+		assert.strictEqual(this.timer.started.length, 0, "Should be '0'");
+		assert.equal(this.timer.stopped.length, 0, "Should be '0'");
+	});
+
+	it("It shouldn't be empty", function () {
 		this.timer.start();
-		test.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
-		test.strictEqual(this.timer.stopped.length, 0, "Should be '0'");
-		test.throws(function () { this.timer.start(); }, Error, "Should be an 'Error'");
-		test.done();
-	}
-};
+		assert.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
+		assert.strictEqual(this.timer.stopped.length, 0, "Should be '0'");
+		assert.throws(function () { this.timer.start(); }, Error, "Should be an 'Error'");
+	});
+});
 
-exports["stop"] = {
-	setUp: function (done) {
+describe("Testing stopping state", function () {
+	beforeEach(function () {
 		this.timer = precise();
-		done();
-	},
-	tests: function (test) {
-		test.expect(4);
-		test.throws(function () { this.timer.stop(); }, Error, "Should be an 'Error'");
-		this.timer.start().stop();
-		test.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
-		test.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
-		test.throws(function () { this.timer.stop(); }, Error, "Should be an 'Error'");
-		test.done();
-	}
-};
+	});
 
-exports["diff"] = {
-	setUp: function (done) {
-		this.timer = precise();
-		done();
-	},
-	tests: function (test) {
-		test.expect(2);
-		test.throws(function () { this.timer.diff(); }, Error, "Should be an 'Error'");
+	it("It shouldn't be empty", function () {
+		assert.throws(function () { this.timer.stop(); }, Error, "Should be an 'Error'");
 		this.timer.start().stop();
-		test.equal(typeof this.timer.diff(), "number", "Shouldn't be 'number'");
-		test.done();
-	}
-};
+		assert.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
+		assert.notEqual(this.timer.started.length, 0, "Shouldn't be '0'");
+		assert.throws(function () { this.timer.stop(); }, Error, "Should be an 'Error'");
+	});
+});
+
+describe("Testing difference", function () {
+	beforeEach(function () {
+		this.timer = precise();
+	});
+
+	it("It should return an integer", function () {
+		assert.throws(function () { this.timer.diff(); }, Error, "Should be an 'Error'");
+		this.timer.start().stop();
+		assert.equal(typeof this.timer.diff(), "number", "Should be 'number'");
+		assert.equal(typeof this.timer.diff(true), "number", "Should be 'number'");
+		assert.equal(this.timer.diff() > this.timer.diff(true), true, "Should be 'true'");
+		assert.equal(parseInt(this.timer.diff() / 1e6, 10), this.timer.diff(true), "Should be 'true'");
+	});
+});
